requirements.txt
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
distro==1.9.0
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
openai==1.58.1
outcome==1.3.0.post0
packaging==24.2
pydantic==2.10.4
pydantic_core==2.27.2
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
